refactor(web): extract useChatOperations from useChatWithSync
Move chat CRUD out of useChatWithSync into a useChatOperations hook:
startNewChat, renameChat, updateChatRepo, updateChatById,
updateCurrentChat, removeChat. It also owns the update/delete/
sandbox-delete mutations (used nowhere else); createChatMutation is
passed in since the draft path shares the same instance.

Public return shape unchanged. Behavior-preserving: logic moved
verbatim, dependency arrays kept. useChatWithSync 724 -> 591.
