+++ packages/web/lib/hooks/useChatWithSync.ts
@@ -13,35 +13,27 @@ import { useSession } from "next-auth/react"
 import { useQueryClient } from "@tanstack/react-query"
 import { nanoid } from "nanoid"
 import type { Chat, ChatStatus, Message } from "@/lib/types"
-import { NEW_REPOSITORY, getDefaultModelForAgent } from "@/lib/types"
+import { getDefaultModelForAgent } from "@/lib/types"
 import type { Credentials } from "@/lib/credentials"
-import {
-  clearLocalStateForChats,
-  collectDescendantIds,
-  DEFAULT_SETTINGS,
-} from "@/lib/storage"
+import { DEFAULT_SETTINGS } from "@/lib/storage"
 import { useChatSyncStore } from "@/lib/stores/chat-sync-store"
 import {
   useChatsQuery,
   useSettingsQuery,
   type SettingsData,
   useCreateChatMutation,
-  useUpdateChatMutation,
-  useDeleteChatMutation,
   useUpdateSettingsMutation,
   useSuggestNameMutation,
-  useSandboxDeleteMutation,
   queryKeys,
 } from "@/lib/query"
 import { useStreamStore } from "@/lib/stores/stream-store"
 import { useStreaming } from "./useStreaming"
 import { useQueueDispatch } from "./useQueueDispatch"
 import { useChatMessageSync } from "./useChatMessageSync"
+import { useChatOperations } from "./useChatOperations"
 import {
   mergeLocalState,
   computeUnseenTransitions,
-  removeLocalChatStateFor,
-  selectFallbackNextChatId,
 } from "@/lib/chat-state"
 import {
   sendMessageToApi,
@@ -70,11 +62,8 @@ export function useChatWithSync() {
 
   // Mutations
   const createChatMutation = useCreateChatMutation()
-  const updateChatMutation = useUpdateChatMutation()
-  const deleteChatMutation = useDeleteChatMutation()
   const updateSettingsMutation = useUpdateSettingsMutation()
   const suggestNameMutation = useSuggestNameMutation()
-  const sandboxDeleteMutation = useSandboxDeleteMutation()
 
   // Local-only state (owned by the chat-sync store)
   const currentChatId = useChatSyncStore((s) => s.currentChatId)
@@ -215,110 +204,15 @@ export function useChatWithSync() {
     }
   }, [createChatMutation])
 
-  // Chat operations
-  const startNewChat = useCallback(async (
-    repo: string = NEW_REPOSITORY,
-    baseBranch: string = "main",
-    parentChatId?: string,
-    switchTo: boolean = true,
-    initialStatus: Chat["status"] = "pending",
-    agent?: string | null,
-    model?: string | null,
-  ): Promise<string | null> => {
-    // Branch chats (with parentChatId) are created immediately since they need to reference the parent
-    if (parentChatId) {
-      try {
-        const newChat = await createChatMutation.mutateAsync({
-          repo,
-          baseBranch,
-          parentChatId,
-          agent,
-          model,
-          status: initialStatus,
-        })
-        if (switchTo) {
-          useChatSyncStore.getState().setCurrentChatId(newChat.id)
-        }
-        return newChat.id
-      } catch (error) {
-        console.error("Failed to create chat:", error)
-        return null
-      }
-    }
-
-    // For regular new chats, enter draft mode instead of creating in DB
-    return useChatSyncStore.getState().enterDraftMode(repo, baseBranch, agent ?? null, model ?? null)
-  }, [createChatMutation])
-
-  const removeChat = useCallback(
-    async (chatId: string, getNextChatId?: (deletedIds: string[]) => string | null) => {
-      const allIds = collectDescendantIds(chats, chatId)
-      for (const id of allIds) useStreamStore.getState().stopStream(id)
-      useChatSyncStore.getState().addDeleting(allIds)
-
-      const selectNextChat = (deletedIds: string[]) => {
-        const nextChat = getNextChatId
-          ? getNextChatId(deletedIds)
-          : selectFallbackNextChatId(chats, deletedIds)
-        useChatSyncStore.getState().setCurrentChatId(nextChat)
-      }
-
-      // Select the next chat right away (optimistically) when the open chat is
-      // being deleted, so the UI moves off it immediately instead of lingering
-      // until the server round-trip completes. The sidebar already removes the
-      // chat optimistically via the delete mutation's onMutate.
-      if (allIds.includes(currentChatId ?? "")) {
-        selectNextChat(allIds)
-      }
-
-      try {
-        const result = await deleteChatMutation.mutateAsync(chatId)
-        for (const sandboxId of result.sandboxIdsToCleanup) {
-          sandboxDeleteMutation.mutate(sandboxId)
-        }
-        clearLocalStateForChats(result.deletedChatIds)
-        useChatSyncStore.getState().setLocalChatState((prev) => removeLocalChatStateFor(prev, result.deletedChatIds))
-        // Reconcile against the server's actual deleted set in case it removed
-        // descendants we didn't predict locally and the open chat was among them.
-        const serverDeletedExtra = result.deletedChatIds.some((id) => !allIds.includes(id))
-        if (serverDeletedExtra && result.deletedChatIds.includes(currentChatId ?? "")) {
-          selectNextChat(result.deletedChatIds)
-        }
-      } catch (error) {
-        console.error("Failed to delete chat:", error)
-      } finally {
-        useChatSyncStore.getState().removeDeleting(allIds)
-      }
-    },
-    [chats, currentChatId, deleteChatMutation, sandboxDeleteMutation]
-  )
-
-  const renameChat = useCallback(async (chatId: string, newName: string) => {
-    try {
-      await updateChatMutation.mutateAsync({ chatId, data: { displayName: newName } })
-    } catch (error) {
-      console.error("Failed to rename chat:", error)
-    }
-  }, [updateChatMutation])
-
-  const updateChatRepo = useCallback(async (chatId: string, repo: string, baseBranch: string) => {
-    const chat = chats.find((c) => c.id === chatId)
-    if (!chat) return
-    // Can select existing repo only before first message and sandbox creation
-    const canSelectExistingRepo = chat.messages.length === 0 && !chat.sandboxId
-    // Can assign a new repo if chat currently has NEW_REPOSITORY
-    const canAssignNewRepo = chat.repo === NEW_REPOSITORY && repo !== NEW_REPOSITORY
-    if (!canSelectExistingRepo && !canAssignNewRepo) return
-
-    try {
-      // When assigning a new repo to an existing sandbox, preserve the working branch.
-      // Only reset branch to null when selecting a repo before sandbox creation.
-      const branchToSet = canAssignNewRepo ? chat.branch : null
-      await updateChatMutation.mutateAsync({ chatId, data: { repo, baseBranch, branch: branchToSet } })
-    } catch (error) {
-      console.error("Failed to update chat repo:", error)
-    }
-  }, [chats, updateChatMutation])
+  // Chat CRUD (create/rename/repo/generic-update/delete) lives in its own hook.
+  const {
+    startNewChat,
+    renameChat,
+    updateChatRepo,
+    updateChatById,
+    updateCurrentChat,
+    removeChat,
+  } = useChatOperations({ chats, currentChatId, createChatMutation })
 
   const updateSettings = useCallback(async (data: { settings?: Partial<typeof settings>; credentials?: Credentials }): Promise<{ ok: boolean; error?: string }> => {
     try {
@@ -328,33 +222,6 @@ export function useChatWithSync() {
       return { ok: false, error: error instanceof Error ? error.message : "Failed to save settings" }
     }
   }, [updateSettingsMutation])
-
-  // Split a Partial<Chat> into the local-only preview fields (handled by the
-  // sync store) and the server-bound fields (sent through the mutation). The
-  // `in`-check matters: destructuring would produce `undefined` whether the
-  // key was present or absent, and an all-undefined preview update is the
-  // store's "clear it" sentinel — so an unrelated update like { planModeEnabled:
-  // false } would otherwise wipe the preview pane.
-  const updateChatById = useCallback(async (chatId: string, updates: Partial<Chat>) => {
-    const { previewItems, activePreviewIndex, previewPaneHidden, queuedMessages, queuePaused, ...serverUpdates } = updates
-
-    if ("previewItems" in updates || "activePreviewIndex" in updates || "previewPaneHidden" in updates) {
-      useChatSyncStore.getState().setPreviewStateForChat(chatId, { previewItems, activePreviewIndex, previewPaneHidden })
-    }
-
-    if (Object.keys(serverUpdates).length > 0) {
-      try {
-        await updateChatMutation.mutateAsync({ chatId, data: serverUpdates as Parameters<typeof updateChatMutation.mutateAsync>[0]["data"] })
-      } catch (error) {
-        console.error("Failed to update chat:", error)
-      }
-    }
-  }, [updateChatMutation])
-
-  const updateCurrentChat = useCallback(async (updates: Partial<Chat>) => {
-    if (!currentChatId) return
-    await updateChatById(currentChatId, updates)
-  }, [currentChatId, updateChatById])
 
   // Send message
   const sendMessage = useCallback(async (content: string, agent?: string, model?: string, files?: File[], targetChatId?: string, planMode?: boolean) => {
